refactor: Command, Query Service 분리 (#7)
