+++ backend/app/src/main/java/carrot/candy/app/chat/controller/ChatRoomController.java
@@ -3,7 +3,8 @@ package carrot.candy.app.chat.controller;
 import carrot.candy.app.auth.domain.AuthMember;
 import carrot.candy.app.chat.dto.request.ChatRoomCreateRequest;
 import carrot.candy.app.chat.dto.response.ChatRoomResponse;
-import carrot.candy.app.chat.service.ChatRoomService;
+import carrot.candy.app.chat.service.ChatRoomCommandService;
+import carrot.candy.app.chat.service.ChatRoomQueryService;
 import carrot.candy.app.common.annotation.PreAuthorize;
 import java.net.URI;
 import java.util.List;
@@ -20,14 +21,15 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/v1/chat-rooms")
 public class ChatRoomController {
 
-    private final ChatRoomService chatRoomService;
+    private final ChatRoomCommandService chatRoomCommandService;
+    private final ChatRoomQueryService chatRoomQueryService;
 
     @PostMapping
     @PreAuthorize
     public ResponseEntity<Void> create(
             AuthMember authMember,
             @RequestBody ChatRoomCreateRequest request) {
-        Long id = chatRoomService.createChatRoom(authMember, request);
+        Long id = chatRoomCommandService.createChatRoom(authMember, request);
         return ResponseEntity.created(URI.create("/api/v1/chat-rooms/" + id))
                 .build();
     }
@@ -35,7 +37,7 @@ public class ChatRoomController {
     @GetMapping
     @PreAuthorize
     public ResponseEntity<List<ChatRoomResponse>> findAll(AuthMember authMember) {
-        List<ChatRoomResponse> chatRooms = chatRoomService.findAllChatRoom(authMember);
+        List<ChatRoomResponse> chatRooms = chatRoomQueryService.findAllChatRoom(authMember);
         return ResponseEntity.ok(chatRooms);
     }
 }
