+++ backend/app/src/main/java/carrot/candy/app/chat/service/ChatRoomCommandService.java
@@ -4,16 +4,14 @@ import carrot.candy.app.auth.domain.AuthMember;
 import carrot.candy.app.chat.domain.chatroom.ChatRoom;
 import carrot.candy.app.chat.domain.chatroom.ChatRoomRepository;
 import carrot.candy.app.chat.dto.request.ChatRoomCreateRequest;
-import carrot.candy.app.chat.dto.response.ChatRoomResponse;
 import carrot.candy.app.member.domain.Member;
 import carrot.candy.app.member.domain.MemberRepository;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class ChatRoomService {
+public class ChatRoomCommandService {
 
     private final ChatRoomRepository chatRoomRepository;
     private final MemberRepository memberRepository;
@@ -24,19 +22,6 @@ public class ChatRoomService {
         validateCreateChatRoom(owner, visitor);
         return chatRoomRepository.save(ChatRoom.createChatRoom(request.name(), visitor, owner))
                 .getId();
-    }
-
-    public List<ChatRoomResponse> findAllChatRoom(AuthMember authMember) {
-        Long memberId = authMember.getId();
-        List<ChatRoom> chatRooms = chatRoomRepository.findAllByOwnerOrVisitorId(memberId);
-        return chatRooms.stream()
-                .map(chatRoom -> {
-                    if (chatRoom.getVisitor().getId() == memberId) {
-                        return ChatRoomResponse.of(chatRoom.getId(), chatRoom.getOwner());
-                    }
-                    return ChatRoomResponse.of(chatRoom.getId(), chatRoom.getVisitor());
-                })
-                .toList();
     }
 
     private void validateCreateChatRoom(Member owner, Member visitor) {
