+++ backend/app/src/main/java/carrot/candy/app/chat/service/MessageCommandService.java
@@ -12,7 +12,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
-public class MessageService {
+public class MessageCommandService {
 
     private final MessageRepository messageRepository;
     private final ChatRoomRepository chatRoomRepository;
